requirements.txt
datasets
transformers
torch
sentencepiece
evaluate
tqdm
pyyaml